Lab #9-answers (Worked with Brandon Krakora)
Were the resulting word frequencies any different?
RESULTS:

MyHashMap result:

time in ms: 4012
the=33194
of=17956
and=12614
a=12346
to=12233
in=9232
was=7542
that=6278
he=5955
his=5496

TreeMap result:

time in ms: 4395
10 most frequent words and their frequencies are basically dublicates of those above.

HashMap result:

Time in ms: 3351
10 most frequent words and their frequencies are same as above.  (like precious answer)

Is the time performance any different? If so, how would you rank the three implementations (in increasing order of time complexity)?

Yes. It is in increasing order in regards to time complexity: HashMap < MyHashMap < TreeMap

How are % and Math.floorMod different? Which works more reliably for computing a hash table index?

The two have no difference when running results on TreeMap and HashMap, but throws Exception on
MyHashMap:
Exception in thread "main" java.lang.IndexOutOfBoundsException: Index -2994 out-of-bounds for length 6007

Also, Math.floorMod performs better than %, it saves almost 1/3 of time on TreeMap.
Therefore, Math.floorMod appears to be more reliable.

What is the time complexity of MyHashMap.size(), and how could you make it much more efficient?

The time complexity of MyHashMap.size() is O(n). I used a private int numKeys to count the chains and return numKeys to make it as O(1).

How does this implementation compare to one where you would directly use your linked Node class from the earlier assignment?Answer briefly in terms of ease of implementation, correctness, reliability, and performance.

Implementation:
This particular implementation uses Iterator to traverse the list, and Node uses its link to next Node and index of the elements. However, HashMap implements circular list, which is better than Node in remove or add/put, because there is no need to define the method at head, tail or middle.

Correctness:
I think the linked Node class is superior because of the possible collision(s) on HashMap.

Reliability:
It seems that HashMap is better since Node throws too many Exceptions.

Performance:
For this one, in my experience, HashMap is better performance wise, with the complexity of most methods is O(1).